根据需求删除白线，提高精度 从int改为float 从而解决填不满的bug

+++ app/src/main/java/com/xiaoziqianbao/piechartview/PiePartBean.java
@@ -11,7 +11,7 @@ public class PiePartBean {
     int endPointX;//扇形末端的点 用来链接圆心画直线
     int endPointY;//扇形末端的点 用来链接圆心画直线
     int radius;//扇形所对圆的半径
-    public  int startArc;//
-    public int moveArc;//划过的角度
+    public  float startArc;//
+    public float moveArc;//划过的角度
     public boolean drawLine = false;
 }
